Add OverrideTool option and refactor ModEntry
- Introduced `OverrideTool` property in `ModConfig` to allow players to override tool usage on chests.
- Updated `ConfigMenu` to register the new option with the Generic Mod Config Menu.
- Refactored `ModEntry` for better configuration initialization and event handling.
- Moved Harmony patches to a new `ModPatches` class for improved organization.
- Removed `AlphaNumeric` string from `Constants` class for code cleanup.
- Updated localization files to include entries for the new `Override Tool` option.
- Modified `CarryChests.sln.DotSettings` to add a new resource dictionary entry.
- Enhanced overall code structure for readability and maintainability.
