+++ CarryChests/i18n/vi.json
@@ -2,6 +2,8 @@
   "alert.carry-chest-limit": "Cannot carry any more chests",
   "config-option.open-held-chest.name": "Open Held Chests",
   "config-option.open-held-chest.description": "Allow you to open the currently held chest.",
+  "config-option.override-toool.name": "Override Tool",
+  "config-option.override-tool.description": "Override using a tool on a chest to pick it up.",
   "config-option.total-limit.name": "Carry Chest Limit",
   "config-option.total-limit.description": "Limits the number of chests that can be carried at once.",
   "config-option.slowness-amount.name": "Slowness Amount",
